added formulas for health/power/speed, added reversing algorithm to obtain base stats for a pet directly from WoW api

+++ src/com/jack/wow/files/api/ApiStats.java
@@ -6,7 +6,7 @@ public class ApiStats
   public int breedId;
   public int petQualityId;
   public int level;
-  public int health;
-  public int power;
-  public int speed;
+  public float health;
+  public float power;
+  public float speed;
 }
